+++ src/rules/standalone/no-browsertestingmodule.ts
@@ -1,0 +1,33 @@
+import type { RuleDefinition } from "@eslint/core";
+import type { TSESTree } from "@typescript-eslint/utils";
+import { isImportIdentifier } from "../../utils/is-import-identifier";
+
+export const ruleName = "no-browsertestingmodule";
+const messageId = "noBrowsertestingmodule";
+
+export const ruleDefinition: RuleDefinition = {
+  meta: {
+    type: "problem",
+    messages: {
+      [messageId]: `\`BrowserTestingModule\` is useless in a standalone application.`,
+    },
+    docs: {
+      description: `Checks that \`BrowserTestingModule\` is not used.`,
+      url: 'https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/standalone/NO_BROWSERTESTINGMODULE.md',
+      recommended: true,
+    },
+    schema: [],
+  },
+  create(context) {
+    return {
+      "Identifier[name='BrowserTestingModule']"(node: TSESTree.Identifier) {
+        if (!isImportIdentifier(node)) {
+          context.report({
+            node,
+            messageId,
+          });
+        }
+      },
+    };
+  },
+};