+++ src/rules/standalone/no-serviceworkermodule.ts
@@ -1,0 +1,33 @@
+import type { RuleDefinition } from "@eslint/core";
+import type { TSESTree } from "@typescript-eslint/utils";
+import { isImportIdentifier } from "../../utils/is-import-identifier";
+
+export const ruleName = "no-serviceworkermodule";
+const messageId = "noServiceworkermodule";
+
+export const ruleDefinition: RuleDefinition = {
+  meta: {
+    type: "problem",
+    messages: {
+      [messageId]: `\`ServiceWorkerModule\` should be avoided in a standalone application, use \`provideServiceWorker()\` instead.`,
+    },
+    docs: {
+      description: `Checks that \`ServiceWorkerModule\` is not used.`,
+      url: 'https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/standalone/NO_SERVICEWORKERMODULE.md',
+      recommended: true,
+    },
+    schema: [],
+  },
+  create(context) {
+    return {
+      "Identifier[name='ServiceWorkerModule']"(node: TSESTree.Identifier) {
+        if (!isImportIdentifier(node)) {
+          context.report({
+            node,
+            messageId,
+          });
+        }
+      },
+    };
+  },
+};