+++ src/rules/standalone/no-withmodule-testing.ts
@@ -1,0 +1,30 @@
+import type { RuleDefinition } from "@eslint/core";
+import type { TSESTree } from "@typescript-eslint/utils";
+
+export const ruleName = "no-withmodule-testing";
+const messageId = "noWithmoduleTesting";
+
+export const ruleDefinition: RuleDefinition = {
+  meta: {
+    type: "problem",
+    messages: {
+      [messageId]: `\`withModule()\` should be avoided in a standalone application, use \`TestBed.configureTestingModule()\` instead.`,
+    },
+    docs: {
+      description: `Checks that \`withModule()\` is not used.`,
+      url: 'https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/standalone/NO_WITHMODULE_TESTING.md',
+      recommended: true,
+    },
+    schema: [],
+  },
+  create(context) {
+    return {
+      "CallExpression[callee.type='Identifier'][callee.name='withModule']"(node: TSESTree.CallExpression) {
+        context.report({
+          node,
+          messageId,
+        });
+      },
+    };
+  },
+};